backend/requirements.txt
# HRMS Backend Requirements
# Hybrid: Vercel Postgres (Prod) + SQLite (Dev)

fastapi>=0.110.0
uvicorn>=0.25.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic>=2.0.0
email-validator>=2.0.0
starlette>=0.37.0
python-dotenv>=1.0.0
